feat: add confirmation dialog before deleting the client + some btn improvements

+++ src/app/pages/list/list.component.ts
@@ -1,23 +1,35 @@
 import { AfterViewInit, Component, inject, OnInit, ViewChild } from '@angular/core';
+import { MatButtonModule } from '@angular/material/button';
+import { MatDialog } from '@angular/material/dialog';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatIconModule } from '@angular/material/icon';
 import { MatInputModule } from '@angular/material/input';
 import { MatSort, MatSortModule } from '@angular/material/sort';
 import { MatTableDataSource, MatTableModule } from '@angular/material/table';
 import { Router } from '@angular/router';
+import { ConfirmationDialog } from '../../components/confirmation-dialog/confiramtion-dialog.component';
 import { Client } from '../../models/client.model';
 import { ApiService } from '../../services/api.service';
 import { InterestsNamesPipe } from '../../shared/pipes/interests-names.pipe';
 
 @Component({
   selector: 'app-list',
-  imports: [MatTableModule, MatIconModule, MatFormFieldModule, MatSortModule, MatInputModule, InterestsNamesPipe],
+  imports: [
+    MatTableModule,
+    MatIconModule,
+    MatFormFieldModule,
+    MatSortModule,
+    MatInputModule,
+    InterestsNamesPipe,
+    MatButtonModule,
+  ],
   templateUrl: './list.component.html',
   styleUrl: './list.component.scss',
 })
 export class ListComponent implements OnInit, AfterViewInit {
   private apiService = inject(ApiService);
   private router = inject(Router);
+  private dialog = inject(MatDialog);
   clients$ = this.apiService.getClients();
   displayedColumns = ['index', 'name', 'surname', 'phone', 'interest', 'actions'];
   @ViewChild(MatSort) sort!: MatSort;
@@ -43,11 +55,20 @@ export class ListComponent implements OnInit, AfterViewInit {
     this.router.navigate(['/edit', id]);
   }
 
-  onDelete(id: string) {
-    this.apiService.removeClient(id);
-  }
-
   onAdd() {
     this.router.navigate(['/add']);
+  }
+
+  openDialog(client: Client): void {
+    const dialogRef = this.dialog.open(ConfirmationDialog, {
+      data: client,
+      width: '250px',
+    });
+
+    dialogRef.afterClosed().subscribe(id => {
+      if (!id) return;
+
+      this.apiService.removeClient(id);
+    });
   }
 }
